Added Error handling and validation

+++ controllers/adventurers.js
@@ -3,21 +3,46 @@ const ObjectId = require('mongodb').ObjectId;
 
 const getAll = async (req, res) => {
     //#swagger.tags=['adventurers']
-    const result = await mongodb.getDatabase().db().collection('adventurers').find();
-    result.toArray().then((adventurers) => {
+    try {
+        const result = await mongodb.getDatabase().db().collection('adventurers').find();
+         
+        const adventurers = await result.toArray();
+
         res.setHeader('Content-Type', 'application/json');
         res.status(200).json(adventurers);
-    });
+    } catch (error) {
+        res.status(500).json({
+            message: 'Some error occurred while fetching the adventurers',
+            error: error.message
+        });
+    }
 };
 
 const getSingle = async (req, res) => {
     //#swagger.tags=['adventurers']
+
+    if (!ObjectId.isValid(req.params.id)) {
+        return res.status(400).json("Id is not valid");
+    }
+
     const adventurersId = new ObjectId(req.params.id);
-    const result = await mongodb.getDatabase().db().collection('adventurers').find({ _id: adventurersId});
-    result.toArray().then((adventurers) => {
+
+    try {
+        const result = await mongodb.getDatabase().db().collection('adventurers').find({ _id: adventurersId});
+        const adventurers = await result.toArray();
         res.setHeader('Content-Type', 'application/json');
+
+        if (adventurers.length === 0) {
+            return res.status(404).json("Record does not exist");
+        }
         res.status(200).json(adventurers[0]);
-    })
+
+    } catch (error) {
+        res.status(500).json({
+            message: 'Some error occurred while fetching the adventurers',
+            error: error.message
+        });
+    }
 };
 
 const createAdventurer = async (req, res) => {
@@ -30,19 +55,37 @@ const createAdventurer = async (req, res) => {
         race: req.body.race,
         gold: req.body.gold
     };
-    const response = await mongodb.getDatabase().db().collection('adventurers').insertOne(adventurer);
 
-     if (response.acknowledged > 0) {
-        res.status(204).send();
+    if (!adventurer.name || !adventurer.class || !adventurer.level || !adventurer.experience || !adventurer.race || !adventurer.gold) {
+        return res.status(400).json("All fields are required");
     }
-    else {
-        res.status(400).json(response.error || 'Some error occured while creating the adventurer');
+
+    try {
+        const response = await mongodb.getDatabase().db().collection('adventurers').insertOne(adventurer);
+
+        if (response.acknowledged > 0) {
+            res.status(204).send();
+        }
+        else {
+            res.status(400).json(response.error || 'Some error occured while creating the adventurer');
+        }
+    } catch (error) {
+        res.status(500).json({
+            message: 'Some error occurred while creating the adventurer',
+            error: error.message
+        });
     }
+    
 };
 
 const updateAdventurer= async (req, res) => {
     //#swagger.tags=['adventurers']
+    if (!ObjectId.isValid(req.params.id)) {
+        return res.status(400).json("Id is not valid");
+    }
+
     const adventurerId = new ObjectId(req.params.id);
+
     const adventurer = {
         name: req.body.name,
         class: req.body.class,
@@ -51,27 +94,59 @@ const updateAdventurer= async (req, res) => {
         race: req.body.race,
         gold: req.body.gold
     };
-    const response = await mongodb.getDatabase().db().collection('adventurers').replaceOne({ _id: adventurerId }, adventurer);
 
-    if (response.acknowledged > 0) {
-        res.status(204).send();
+    if (!adventurer.name || !adventurer.class || !adventurer.level || !adventurer.experience || !adventurer.race || !adventurer.gold) {
+        return res.status(400).json("All fields are required");
     }
-    else {
-        res.status(400).json(response.error || 'Some error occured while updating the adventurer');
+
+    try {
+        const response = await mongodb.getDatabase().db().collection('adventurers').replaceOne({ _id: adventurerId }, adventurer);
+        
+        if (response.matchedCount === 0) {
+            return res.status(404).json("Record does not exist");
+        }
+        if (response.acknowledged > 0) {
+            res.status(204).send();
+        }
+        else {
+            res.status(400).json(response.error || 'Some error occured while updating the adventurer');
+        }
+    } catch (error) {
+        res.status(500).json({
+            message: 'Some error occurred while updating the adventurer',
+            error: error.message
+        });
     }
+    
 };
 
 const deleteAdventurer = async (req, res) => {
     //#swagger.tags=['adventurers']
-    const adventurerId = new ObjectId(req.params.id);
-    const response = await mongodb.getDatabase().db().collection('adventurers').deleteOne({ _id: adventurerId }); 
+    if (!ObjectId.isValid(req.params.id)) {
+        return res.status(400).json("Id is not valid");
+    }
 
-    if (response.deletedCount > 0) {
-        res.status(204).send();
+    const adventurerId = new ObjectId(req.params.id);
+
+    try {
+        const response = await mongodb.getDatabase().db().collection('adventurers').deleteOne({ _id: adventurerId }); 
+       
+        if (response.deletedCount === 0) {
+            return res.status(404).json("Record does not exist");
+        }
+        if (response.deletedCount > 0) {
+            res.status(204).send();
+        }
+        else {
+            res.status(400).json(response.error || 'Some error occured while deleting the adventurer');
     }
-    else {
-        res.status(400).json(response.error || 'Some error occured while deleting the adventurer');
+    } catch (error) {
+        res.status(500).json({
+            message: 'Some error occurred while deleting the adventurer',
+            error: error.message
+        });
     }
+    
 };
 
 module.exports = {
